No hacky wordCache
